requirements.txt
Pillow~=9.1.0
pip~=20.0.2
keyring~=18.0.1
setuptools~=45.2.0